Error display

+++ chatabook/graphql/resolvers.tsx
@@ -1,11 +1,8 @@
 import { EventWithUser } from '@/pages/events';
 import { createDate } from '@/scripts/createDate';
 import { Event, User, User_role } from '@prisma/client';
-// import prisma from '../lib/prisma';
 import validator from 'validator';
-import { PrismaClient } from '@prisma/client';
-
-const prisma = new PrismaClient();
+import prisma from '@/lib/prisma';
 
 export const resolvers = {
   Query: {
@@ -96,8 +93,34 @@ export const resolvers = {
         };
       }
     },
-    editEvent: async (_: any, args: Event) => {
+    editEvent: async (_: any, args: Event, context: { userEmail: string }) => {
       try {
+        const eventToEdit = await prisma.event.findUnique({
+          where: {
+            id: Number(args.id),
+          },
+          include: {
+            user: true,
+          },
+        });
+
+        const userEditing = await prisma.user.findUnique({
+          where: {
+            email: context.userEmail,
+          },
+        });
+
+        if (!eventToEdit) {
+          throw new Error('Event not found');
+        }
+
+        if (userEditing && userEditing?.roleId < 3) {
+          throw new Error('Your role is not strong enought to edit events');
+        }
+        if (eventToEdit.user.email !== context.userEmail) {
+          throw new Error('You are unauthorized to edit this event');
+        }
+
         const changeEvent = await prisma.event.update({
           where: {
             id: Number(args.id),
@@ -127,7 +150,33 @@ export const resolvers = {
         };
       }
     },
-    deleteEvent(_: any, args: Event) {
+    deleteEvent: async (_: any, args: Event, context: { userEmail: string }) => {
+      const eventToDelete = await prisma.event.findUnique({
+        where: {
+          id: Number(args.id),
+        },
+        include: {
+          user: true,
+        },
+      });
+
+      const userDeleting = await prisma.user.findUnique({
+        where: {
+          email: context.userEmail,
+        },
+      });
+
+      if (!eventToDelete) {
+        throw new Error('Event not found');
+      }
+
+      if (userDeleting && userDeleting?.roleId < 3) {
+        throw new Error('Your role is not strong enought to delete events');
+      }
+      if (eventToDelete.user.email !== context.userEmail) {
+        throw new Error('You are unauthorized to delete this event');
+      }
+
       return prisma.event.delete({
         where: {
           id: Number(args.id),
@@ -143,17 +192,30 @@ export const resolvers = {
             message: 'Email is not valid',
           };
         }
-        const existingUser = await prisma.user.findFirst({
-          where: {
-            email: args.email,
-          },
-        });
+        const [existingUser, phoneNumberInUse] = await Promise.all([
+          prisma.user.findFirst({
+            where: {
+              email: args.email,
+            },
+          }),
+          prisma.user.findFirst({
+            where: {
+              telephone: args.telephone,
+            },
+          }),
+        ]);
 
         if (existingUser) {
           return {
             code: 400,
             success: false,
             message: 'Daný email sa už používa.',
+          };
+        } else if (phoneNumberInUse) {
+          return {
+            code: 400,
+            success: false,
+            message: 'Telefónne číslo sa už používa.',
           };
         }
